Add ignore android flag in performance README.md
The tools/run_tests/performance/README.md provides an example to run workers in Java.  However, this example does not work out of the box, because the Java build requires an Android SDK. This change adds a Gradle flag `-PskipAndroid=true` to the existing example, allowing it to work as intended.